chore(typescript): improved organization of types; changed CommonProps to a named export
