+++ packages/victory-core/src/index.ts
@@ -6,6 +6,7 @@ export * from "./victory-container/victory-container";
 export * from "./victory-label/victory-label";
 export * from "./victory-transition/victory-transition";
 export * from "./victory-clip-container/victory-clip-container";
+export * from "./victory-theme/types";
 export * from "./victory-theme/victory-theme";
 export * from "./victory-portal/portal";
 export * from "./victory-portal/portal-context";
@@ -31,7 +32,7 @@ export * as Style from "./victory-util/style";
 export * as TextSize from "./victory-util/textsize";
 export * as Transitions from "./victory-util/transitions";
 export * as UserProps from "./victory-util/user-props";
-export * as CommonProps from "./victory-util/common-props";
+export * from "./victory-util/common-props";
 export * as Wrapper from "./victory-util/wrapper";
 export * as Axis from "./victory-util/axis";
 export * as Hooks from "./victory-util/hooks";
@@ -39,4 +40,4 @@ export * as LineHelpers from "./victory-util/line-helpers";
 export { default as Timer } from "./victory-util/timer";
 export * as PointPathHelpers from "./victory-util/point-path-helpers";
 export { default as TimerContext } from "./victory-util/timer-context";
-export * from "./victory-util/types";
+export * from "./types/prop-types";
