+++ packages/victory-core/src/victory-primitives/border.tsx
@@ -2,9 +2,11 @@ import React from "react";
 import PropTypes from "prop-types";
 import * as Helpers from "../victory-util/helpers";
 import { assign } from "lodash";
-import * as CommonProps from "../victory-util/common-props";
+import {
+  CommonProps,
+  VictoryCommonPrimitiveProps
+} from "../victory-util/common-props";
 import { Rect } from "./rect";
-import { VictoryCommonPrimitiveProps } from "../victory-util/types";
 
 export interface BorderProps extends VictoryCommonPrimitiveProps {
   width?: number;
