+++ packages/victory-core/src/victory-primitives/tspan.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { VictoryCommonPrimitiveProps } from "../victory-util/types";
+import { VictoryCommonPrimitiveProps } from "../victory-util/common-props";
 
 export const TSpan = (props: VictoryCommonPrimitiveProps) => (
   // @ts-expect-error FIXME: "id cannot be a number"
