+++ packages/victory-core/src/victory-primitives/whisker.tsx
@@ -2,9 +2,11 @@ import React from "react";
 import { assign } from "lodash";
 import PropTypes from "prop-types";
 import * as Helpers from "../victory-util/helpers";
-import * as CommonProps from "../victory-util/common-props";
+import {
+  CommonProps,
+  VictoryCommonPrimitiveProps
+} from "../victory-util/common-props";
 import { Line } from "./line";
-import { VictoryCommonPrimitiveProps } from "../victory-util/types";
 
 export type WhiskerAxes = {
   x1?: number;
