+++ packages/victory-core/src/victory-theme/victory-theme.tsx
@@ -1,5 +1,5 @@
 import { grayscale } from "./grayscale";
 import { material } from "./material";
-export * from "./victory-theme-definition";
+export * from "./types";
 
 export const VictoryTheme = { grayscale, material };
