+++ packages/victory-core/src/victory-util/common-props.tsx
@@ -1,7 +1,39 @@
 import PropTypes from "prop-types";
 import * as CustomPropTypes from "./prop-types";
+import * as React from "react";
+import { OriginType } from "../victory-label/victory-label";
+import { PaddingProps, VictoryThemeDefinition } from "../victory-theme/types";
+import {
+  AnimatePropTypeInterface,
+  CategoryPropType,
+  ColorScalePropType,
+  D3Scale,
+  DataGetterPropType,
+  DomainPaddingPropType,
+  DomainPropType,
+  EventCallbackInterface,
+  RangePropType,
+  ScalePropType,
+  SortOrderPropType,
+  StringOrNumberOrList
+} from "../types/prop-types";
+import { NumberOrCallback, StringOrCallback } from "../types/callbacks";
 
-export const dataProps = {
+export interface VictoryDatableProps {
+  categories?: CategoryPropType;
+  data?: any[];
+  dataComponent?: React.ReactElement;
+  domain?: DomainPropType;
+  domainPadding?: DomainPaddingPropType;
+  samples?: number;
+  sortKey?: DataGetterPropType;
+  sortOrder?: SortOrderPropType;
+  x?: DataGetterPropType;
+  y?: DataGetterPropType;
+  y0?: DataGetterPropType;
+}
+const dataProps: React.WeakValidationMap<VictoryDatableProps> = {
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   categories: PropTypes.oneOfType([
     PropTypes.arrayOf(PropTypes.string),
     PropTypes.shape({
@@ -9,6 +41,7 @@ export const dataProps = {
       y: PropTypes.arrayOf(PropTypes.string)
     })
   ]),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   data: PropTypes.oneOfType([PropTypes.array, PropTypes.object]),
   dataComponent: PropTypes.element,
   disableInlineStyles: PropTypes.bool,
@@ -59,13 +92,49 @@ export const dataProps = {
   ])
 };
 
-export const baseProps = {
+export interface VictoryCommonThemeProps {
+  animate?: boolean | AnimatePropTypeInterface;
+  colorScale?: ColorScalePropType;
+  containerComponent?: React.ReactElement;
+  domainPadding?: DomainPaddingPropType;
+  externalEventMutations?: EventCallbackInterface<
+    string | string[],
+    StringOrNumberOrList
+  >[];
+  groupComponent?: React.ReactElement;
+  height?: number;
+  horizontal?: boolean;
+  maxDomain?: number | { x?: number; y?: number };
+  minDomain?: number | { x?: number; y?: number };
+  name?: string;
+  origin?: OriginType;
+  padding?: PaddingProps;
+  polar?: boolean;
+  range?: RangePropType;
+  scale?:
+    | ScalePropType
+    | D3Scale
+    | {
+        x?: ScalePropType | D3Scale;
+        y?: ScalePropType | D3Scale;
+      };
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  sharedEvents?: { events: any[]; getEventState: Function };
+  singleQuadrantDomainPadding?: boolean | { x?: boolean; y?: boolean };
+  standalone?: boolean;
+  width?: number;
+}
+export interface VictoryCommonProps extends VictoryCommonThemeProps {
+  theme?: VictoryThemeDefinition;
+}
+const baseProps: React.WeakValidationMap<VictoryCommonProps> = {
   animate: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),
   containerComponent: PropTypes.element,
   domain: PropTypes.oneOfType([
     CustomPropTypes.domain,
     PropTypes.shape({ x: CustomPropTypes.domain, y: CustomPropTypes.domain })
   ]),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   maxDomain: PropTypes.oneOfType([
     PropTypes.number,
     PropTypes.instanceOf(Date),
@@ -74,6 +143,7 @@ export const baseProps = {
       y: PropTypes.oneOfType([PropTypes.number, PropTypes.instanceOf(Date)])
     })
   ]),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   minDomain: PropTypes.oneOfType([
     PropTypes.number,
     PropTypes.instanceOf(Date),
@@ -82,6 +152,7 @@ export const baseProps = {
       y: PropTypes.oneOfType([PropTypes.number, PropTypes.instanceOf(Date)])
     })
   ]),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   domainPadding: PropTypes.oneOfType([
     PropTypes.shape({
       x: PropTypes.oneOfType([
@@ -118,6 +189,7 @@ export const baseProps = {
       eventHandlers: PropTypes.object
     })
   ),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   externalEventMutations: PropTypes.arrayOf(
     PropTypes.shape({
       callback: PropTypes.func,
@@ -137,7 +209,9 @@ export const baseProps = {
   groupComponent: PropTypes.element,
   height: CustomPropTypes.nonNegative,
   name: PropTypes.string,
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   origin: PropTypes.shape({ x: PropTypes.number, y: PropTypes.number }),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   padding: PropTypes.oneOfType([
     PropTypes.number,
     PropTypes.shape({
@@ -156,10 +230,12 @@ export const baseProps = {
     CustomPropTypes.scale,
     PropTypes.shape({ x: CustomPropTypes.scale, y: CustomPropTypes.scale })
   ]),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   sharedEvents: PropTypes.shape({
     events: PropTypes.array,
     getEventState: PropTypes.func
   }),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   singleQuadrantDomainPadding: PropTypes.oneOfType([
     PropTypes.bool,
     PropTypes.shape({
@@ -172,7 +248,29 @@ export const baseProps = {
   width: CustomPropTypes.nonNegative
 };
 
-export const primitiveProps = {
+export interface VictoryCommonPrimitiveProps {
+  active?: boolean;
+  ariaLabel?: StringOrCallback;
+  className?: string;
+  clipPath?: string;
+  data?: any;
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  desc?: string | Function;
+  disableInlineStyles?: boolean;
+  events?: object;
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  id?: number | string | Function;
+  index?: number | string;
+  origin?: OriginType;
+  polar?: boolean;
+  role?: string;
+  scale?: any;
+  shapeRendering?: string;
+  style?: any;
+  tabIndex?: NumberOrCallback;
+  transform?: string;
+}
+const primitiveProps: React.WeakValidationMap<VictoryCommonPrimitiveProps> = {
   active: PropTypes.bool,
   ariaLabel: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
   className: PropTypes.string,
@@ -183,6 +281,7 @@ export const primitiveProps = {
   events: PropTypes.object,
   id: PropTypes.oneOfType([PropTypes.number, PropTypes.string, PropTypes.func]),
   index: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  // @ts-expect-error TODO: synchronize the type with this PropTypes
   origin: PropTypes.shape({ x: PropTypes.number, y: PropTypes.number }),
   polar: PropTypes.bool,
   role: PropTypes.string,
@@ -194,4 +293,10 @@ export const primitiveProps = {
   style: PropTypes.object,
   tabIndex: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
   transform: PropTypes.string
+};
+
+export const CommonProps = {
+  dataProps,
+  baseProps,
+  primitiveProps
 };
