+++ packages/victory-core/src/victory-util/textsize.d.ts
@@ -1,4 +1,11 @@
-import { TextSizeStyleInterface } from "./types";
+export interface TextSizeStyleInterface {
+  angle?: number;
+  characterConstant?: string;
+  fontFamily?: string;
+  fontSize?: number | string;
+  letterSpacing?: string;
+  lineHeight?: number;
+}
 
 export function approximateTextSize(
   text: string,
